Fix y's type

+++ main.c
@@ -4,7 +4,8 @@
 
 // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。
 double myRoot(double x) { // fix return type
-    int y;
+    // fix y's type
+    double y;
     if (x <= 0) {
         exit(1);
     }
